ai-diary/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
werkzeug==2.2.3
vertexai==1.71.1
mysql-connector-python>=8.0.33
cloud-sql-python-connector[pymysql]>=1.11.0
PyMySQL>=1.1.0
google-cloud-firestore>=2.11.1
requests==2.28.2
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0